split ptable projector
